+++ docs/_pages/releases.md
@@ -12,6 +12,7 @@ sidebar:
 ### What's New
 
 * Prevent asserting directly on `AndConstraint` - [#1649](https://github.com/fluentassertions/fluentassertions/pull/1649)
+* Added `WithInnerExceptionExactly` extension method on `Task<ExceptionAssertions<T>>` for easier use with `ThrowAsync` - [#1658](https://github.com/fluentassertions/fluentassertions/pull/1658)
 
 ## 6.0.0